pyproject.toml
[tool.poetry]
name = "photofield-ai"
version = "0.1.0"
description = ""
authors = ["Miha Lunar <mlunar@gmail.com>"]
readme = "README.md"
packages = [{include = "photofield_ai"}]

[tool.poetry.dependencies]
python = "^3.10"
onnxruntime = "^1.12.1"
Pillow = "^9.2.0"
numpy = "^1.23.3"
fastapi = "^0.85.0"
ujson = "^5.5.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
python-multipart = "^0.0.5"
line-profiler = "^3.5.1"
ftfy = "^6.1.1"
regex = "^2022.9.13"
onnx = "^1.12.0"
tqdm = "^4.64.1"

[tool.poetry.group.gpu.dependencies]
onnxruntime-gpu = "^1.12.1"

[tool.poetry.group.compile]
optional = true

[tool.poetry.group.compile.dependencies]
Nuitka = "^1.1.2"
ordered-set = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
